add google to chat completions streaming snapshots
