+++ payloads/scripts/transforms/transforms-chat-completions.test.ts
@@ -8,9 +8,11 @@ import {
   getResponsePath,
   getStreamingResponsePath,
   getFixtureSkipReason,
+} from "./helpers";
+import {
   registerSkippedFixtureTest,
   useTransformTestServer,
-} from "./helpers";
+} from "./vitest-helpers";
 
 const TIMEOUT = 30000;
 const getServer = useTransformTestServer();
