+++ payloads/transforms/google_to_chat-completions/multimodalRequest-streaming.json
@@ -1,0 +1,38 @@
+[
+  {
+    "id": "chatcmpl-DRi6z6lAPf01COZ62T25baqMbXFcs",
+    "object": "chat.completion.chunk",
+    "created": 1775496793,
+    "model": "gpt-5-nano-2025-08-07",
+    "service_tier": "default",
+    "system_fingerprint": null,
+    "choices": [
+      {
+        "index": 0,
+        "delta": {
+          "role": "assistant",
+          "content": "",
+          "refusal": null
+        },
+        "finish_reason": null
+      }
+    ],
+    "obfuscation": "oW"
+  },
+  {
+    "id": "chatcmpl-DRi6z6lAPf01COZ62T25baqMbXFcs",
+    "object": "chat.completion.chunk",
+    "created": 1775496793,
+    "model": "gpt-5-nano-2025-08-07",
+    "service_tier": "default",
+    "system_fingerprint": null,
+    "choices": [
+      {
+        "index": 0,
+        "delta": {},
+        "finish_reason": "length"
+      }
+    ],
+    "obfuscation": "c6ZiIryWO2sI"
+  }
+]
