+++ payloads/transforms/google_to_chat-completions/simpleRequestTruncated-streaming.json
@@ -1,0 +1,38 @@
+[
+  {
+    "id": "chatcmpl-DRi6y76JGZrVYj6Y9gv176kaV8WzX",
+    "object": "chat.completion.chunk",
+    "created": 1775496792,
+    "model": "gpt-5-nano-2025-08-07",
+    "service_tier": "default",
+    "system_fingerprint": null,
+    "choices": [
+      {
+        "index": 0,
+        "delta": {
+          "role": "assistant",
+          "content": "",
+          "refusal": null
+        },
+        "finish_reason": null
+      }
+    ],
+    "obfuscation": "m2"
+  },
+  {
+    "id": "chatcmpl-DRi6y76JGZrVYj6Y9gv176kaV8WzX",
+    "object": "chat.completion.chunk",
+    "created": 1775496792,
+    "model": "gpt-5-nano-2025-08-07",
+    "service_tier": "default",
+    "system_fingerprint": null,
+    "choices": [
+      {
+        "index": 0,
+        "delta": {},
+        "finish_reason": "length"
+      }
+    ],
+    "obfuscation": "xCxiJ7d8zkyU"
+  }
+]
