+++ payloads/transforms/google_to_chat-completions/systemMessageArrayContent-streaming.json
@@ -1,0 +1,38 @@
+[
+  {
+    "id": "chatcmpl-DRi72itxzZrzr6RprqayBddeLQAWR",
+    "object": "chat.completion.chunk",
+    "created": 1775496796,
+    "model": "gpt-5-nano-2025-08-07",
+    "service_tier": "default",
+    "system_fingerprint": null,
+    "choices": [
+      {
+        "index": 0,
+        "delta": {
+          "role": "assistant",
+          "content": "",
+          "refusal": null
+        },
+        "finish_reason": null
+      }
+    ],
+    "obfuscation": "SN"
+  },
+  {
+    "id": "chatcmpl-DRi72itxzZrzr6RprqayBddeLQAWR",
+    "object": "chat.completion.chunk",
+    "created": 1775496796,
+    "model": "gpt-5-nano-2025-08-07",
+    "service_tier": "default",
+    "system_fingerprint": null,
+    "choices": [
+      {
+        "index": 0,
+        "delta": {},
+        "finish_reason": "length"
+      }
+    ],
+    "obfuscation": "1yAFWL0oPxr2"
+  }
+]
